refactor: clean up the folder structure. Fonts.kt to ui.theme. Database.kt and Network.kt to data.
