+++ app/src/main/java/com/example/littlelemon/data/Network.kt
@@ -1,4 +1,4 @@
-package com.example.littlelemon
+package com.example.littlelemon.data
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
