+++ app/src/main/java/com/example/littlelemon/ui/theme/Fonts.kt
@@ -1,8 +1,9 @@
-package com.example.littlelemon
+package com.example.littlelemon.ui.theme
 
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
+import com.example.littlelemon.R
 
 val karlaFontFamily = FontFamily(
     Font(R.font.karla_regular, FontWeight.Normal)
